.claude-worktrees/welding defects/requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
